String and also byte array transmittions work.
The protocol for Strings is: One line, one response.
For bytes, it is more complicated because the contents could be anything.
Thus, in the case of bytes, NO OTHER output of the external program
on stdout is allowed! To know when a byte-encoded message begins
and when it ends, the begin of each message is a 4-byte integer
encoding the length of the message in bytes. After those bytes have
been read, the next message is expected.