Actually, it is possible to receive multiple values for one call.
In the last commits, I said it would be a 1:1 association between
receive() calls and received data, making the returning of Streams
obsolete. This was only true for the binary reader as long as the
byte buffer had the length of 1, which I did for testing.
When the buffer is longer, its easily possible to read multiple
messages. This actually currently causes the program to get stuck
in such cases. Trying to solve.