Added a "external program ready signal".
This signal is a simple string line that must be sent by the
external program to indicate that it is ready for processing. If
this is set - via the Options object - that previous output of the
external program will be ignored. This is helpful for situations
where at the startup of the program some output occurs by some
underlying libraries which we want to ignore.