Fixed an issue where the wait for the ready signal blocked forever.
I had used a BufferedReader to get lines from the external program
InputStream until the ready signal was found. However, the BufferedReader
caused the subsequent algorithm to hang. Weirdly, printing anything
in the external script on stderr after giving the ready signal
fixed the issue. But this is not satisfactory, of course.
I now replaced the buffered reader by single reads from the original
InputStream.